# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data."
          },
          "409": {
            "description": "Email already exists."
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (Invalid Credentials)."
          }
        },
        "summary": "Log in a user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (Invalid or missing token)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/projects/{projectId}/cultural-data": {
      "post": {
        "operationId": "CulturalDataController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCulturalDataDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Datos culturales creados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CulturalDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos."
          },
          "404": {
            "description": "Proyecto o Región referenciada no encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear nuevos datos culturales dentro de un proyecto",
        "tags": [
          "Cultural Data"
        ]
      },
      "get": {
        "operationId": "CulturalDataController_findAll",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de datos culturales.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CulturalDataResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Proyecto no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todos los datos culturales de un proyecto",
        "tags": [
          "Cultural Data"
        ]
      }
    },
    "/api/projects/{projectId}/cultural-data/{culturalDataId}": {
      "get": {
        "operationId": "CulturalDataController_findOne",
        "parameters": [
          {
            "name": "culturalDataId",
            "required": true,
            "in": "path",
            "description": "ID de los datos culturales",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos culturales encontrados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CulturalDataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Proyecto o Datos culturales no encontrados."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener datos culturales por ID dentro de un proyecto",
        "tags": [
          "Cultural Data"
        ]
      },
      "patch": {
        "operationId": "CulturalDataController_update",
        "parameters": [
          {
            "name": "culturalDataId",
            "required": true,
            "in": "path",
            "description": "ID a actualizar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCulturalDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Datos culturales actualizados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CulturalDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos."
          },
          "404": {
            "description": "Proyecto o Datos culturales no encontrados."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar datos culturales por ID dentro de un proyecto",
        "tags": [
          "Cultural Data"
        ]
      },
      "delete": {
        "operationId": "CulturalDataController_remove",
        "parameters": [
          {
            "name": "culturalDataId",
            "required": true,
            "in": "path",
            "description": "ID a eliminar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos culturales eliminados."
          },
          "404": {
            "description": "Proyecto o Datos culturales no encontrados."
          },
          "409": {
            "description": "Conflicto al eliminar (referenciado por otras entidades)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar datos culturales por ID dentro de un proyecto",
        "tags": [
          "Cultural Data"
        ]
      }
    },
...
  },
  "info": {
    "title": "JAPM API",
    "description": "API para la aplicación de Prompt Engineering JAPM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "User password",
            "example": "Password123!"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "description": "User full name",
            "example": "Test User"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "createdAt"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "Password123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ]
      },
      "CreateRegionDto": {
        "type": "object",
        "properties": {
          "languageCode": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "description": "Código de idioma único que actúa como ID",
            "example": "de-DE"
          },
          "name": {
            "type": "string",
            "description": "Nombre de la región",
            "example": "Alemania"
          },
          "parentRegionId": {
            "type": "string",
            "description": "languageCode de la región padre (opcional)",
            "example": "eu"
          },
          "timeZone": {
            "type": "string",
            "description": "Zona horaria",
            "example": "Europe/Berlin"
          },
          "defaultFormalityLevel": {
            "type": "string",
            "description": "Nivel de formalidad por defecto (opcional)",
            "example": "Formal"
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionales (opcional)"
          }
        },
        "required": [
          "languageCode",
          "name"
        ]
      },
...
      "CreateCulturalDataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único para estos datos culturales (formato slug)",
            "example": "direct-and-formal"
          },
          "regionId": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "description": "ID de la región asociada (código de idioma xx-XX)",
            "example": "es-ES"
          },
          "formalityLevel": {
            "type": "number",
            "description": "Nivel de formalidad (opcional)",
            "example": 5
          },
          "style": {
            "type": "string",
            "description": "Estilo de comunicación (opcional)",
            "example": "Directo"
          },
          "considerations": {
            "type": "string",
            "description": "Consideraciones culturales (opcional)"
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionales (opcional)"
          }
        },
        "required": [
          "id",
          "regionId"
        ]
      },
      "CulturalDataResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único para estos datos culturales (formato slug)",
            "example": "direct-and-formal"
          },
          "regionId": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "description": "ID de la región asociada (código de idioma xx-XX)",
            "example": "es-ES"
          },
          "formalityLevel": {
            "type": "number",
            "description": "Nivel de formalidad (opcional)",
            "example": 5
          },
          "style": {
            "type": "string",
            "description": "Estilo de comunicación (opcional)",
            "example": "Directo"
          },
          "considerations": {
            "type": "string",
            "description": "Consideraciones culturales (opcional)"
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionales (opcional)"
          },
          "region": {
            "$ref": "#/components/schemas/CreateRegionDto"
          },
          "projectId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "regionId",
          "region",
          "projectId"
        ]
      },
      "UpdateCulturalDataDto": {
        "type": "object",
        "properties": {
          "formalityLevel": {
            "type": "number",
            "description": "Nivel de formalidad (opcional)",
            "example": 5
          },
          "style": {
            "type": "string",
            "description": "Estilo de comunicación (opcional)",
            "example": "Directo"
          },
          "considerations": {
            "type": "string",
            "description": "Consideraciones culturales (opcional)"
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionales (opcional)"
          }
        }
      },
...
    }
  }
}