# src/auth/dto/login.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEmail, IsString, MinLength } from 'class-validator';

export class LoginDto {
    @ApiProperty({ description: 'User email address', example: 'test@example.com' })
    @IsEmail()
    email: string;

    @ApiProperty({ description: 'User password', example: 'Password123!' })
    @IsString()
    password: string;
} 