# src/auth/dto/register.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEmail, IsString, MinLength, MaxLength } from 'class-validator';

export class RegisterDto {
    @ApiProperty({ description: 'User email address', example: 'test@example.com' })
    @IsEmail()
    email: string;

    @ApiProperty({ description: 'User password', example: 'Password123!', minLength: 8 })
    @IsString()
    @MinLength(8)
    password: string;

    @ApiProperty({ description: 'User full name', example: 'Test User', maxLength: 100 })
    @IsString()
    @MaxLength(100)
    name: string;
} 